requirements.txt
gradio
openai
langchain
pinecone-client
tiktoken